+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/configuration/Configuration.java
@@ -102,6 +102,7 @@ public class Configuration {
     UnaryOperator.identity();
   private UnaryOperator<SslContextBuilder> sslClientContextBuilderCustomizer =
     UnaryOperator.identity();
+  private Function<java.security.cert.X509Certificate, byte[]> ocspResponseSupplier = null;
 
 
   // inbound - dynamic private key & x509