+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyTls.java
@@ -15,6 +15,7 @@ import com.github.tomakehurst.wiremock.junit5.WireMockTest;
 import de.gematik.rbellogger.data.RbelElementAssertion;
 import de.gematik.rbellogger.util.CryptoLoader;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerProxyConfiguration;
+import de.gematik.test.tiger.common.data.config.tigerproxy.TigerProxyConfiguration.TigerProxyConfigurationBuilder;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerRoute;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerTlsConfiguration;
 import de.gematik.test.tiger.common.pki.TigerConfigurationPkiIdentity;
@@ -23,27 +24,20 @@ import de.gematik.test.tiger.config.ResetTigerConfiguration;
 import io.restassured.RestAssured;
 import io.restassured.config.SSLConfig;
 import io.restassured.response.Response;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.OutputStream;
+import java.io.*;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.net.Proxy.Type;
 import java.net.Socket;
 import java.security.*;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateExpiredException;
-import java.security.cert.CertificateNotYetValidException;
-import java.security.cert.X509Certificate;
+import java.security.cert.*;
 import java.time.ZonedDateTime;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.net.ssl.*;
@@ -62,6 +56,8 @@ import org.apache.http.ssl.SSLContexts;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.jsse.provider.BouncyCastleJsseProvider;
+import org.bouncycastle.tls.*;
+import org.bouncycastle.tls.crypto.impl.bc.BcTlsCrypto;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -739,7 +735,7 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
 
               client.newCall(request).execute();
             })
-      .isInstanceOf(SSLHandshakeException.class);
+        .isInstanceOf(SSLHandshakeException.class);
   }
 
   @Test
@@ -784,7 +780,7 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
               RestAssured.proxy("localhost", tigerProxy.getProxyPort());
               RestAssured.get("https://backend/foobar").andReturn();
             })
-      .isInstanceOf(SSLHandshakeException.class);
+        .isInstanceOf(SSLHandshakeException.class);
   }
 
   @Test
@@ -871,5 +867,73 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
 
     assertThat(response.getStatus()).isEqualTo(666);
     assertThat(response.getBody().getObject().get("foo")).hasToString("bar");
+  }
+
+  @Test
+  void tigerProxyAsServerWithActivatedOcspStapling_shouldSendValidOcspResponseInHandshake() {
+    spawnTigerProxyAndConnectWithBouncyCastleAndCheckServerCertificate(
+        serverCertificate ->
+            assertThat(
+                    serverCertificate
+                        .getCertificateStatus()
+                        .getOCSPResponse()
+                        .getResponseStatus()
+                        .getIntValue())
+                .isZero(),
+        TigerProxyConfiguration.builder()
+            .tls(
+                TigerTlsConfiguration.builder()
+                    .ocspSignerIdentity(
+                        new TigerConfigurationPkiIdentity("src/test/resources/ocspSigner.p12;00"))
+                    .build()));
+  }
+
+  @Test
+  void tigerProxyAsServerWithoutActivatedOcspStapling_shouldNotSendValidOcspResponseInHandshake() {
+    spawnTigerProxyAndConnectWithBouncyCastleAndCheckServerCertificate(
+        serverCertificate -> assertThat(serverCertificate.getCertificateStatus()).isNull(),
+        TigerProxyConfiguration.builder().tls(TigerTlsConfiguration.builder().build()));
+  }
+
+  private void spawnTigerProxyAndConnectWithBouncyCastleAndCheckServerCertificate(
+      Consumer<TlsServerCertificate> serverCertificateConsumer,
+      TigerProxyConfigurationBuilder proxyConfigurationBuilder) {
+    AtomicInteger checkCounter = new AtomicInteger(0);
+
+    spawnTigerProxyWith(
+        proxyConfigurationBuilder
+            .proxyRoutes(
+                List.of(
+                    TigerRoute.builder()
+                        .from("/")
+                        .to("http://localhost:" + fakeBackendServerPort)
+                        .build()))
+            .build());
+
+    try (Socket socket = new Socket("127.0.0.1", tigerProxy.getProxyPort())) {
+      TlsClientProtocol tlsClientProtocol =
+          new TlsClientProtocol(socket.getInputStream(), socket.getOutputStream());
+      tlsClientProtocol.connect(
+          new DefaultTlsClient(new BcTlsCrypto()) {
+            @Override
+            public TlsAuthentication getAuthentication() {
+              return new ServerOnlyTlsAuthentication() {
+                @Override
+                public void notifyServerCertificate(TlsServerCertificate serverCertificate) {
+                  serverCertificateConsumer.accept(serverCertificate);
+                  checkCounter.incrementAndGet();
+                }
+              };
+            }
+          });
+      tlsClientProtocol.getOutputStream().write("GET /foobar HTTP/1.1\r\n\r\n".getBytes());
+
+      assertThat(checkCounter)
+          .withFailMessage(
+              () -> "No notification of server certificate. Maybe the connection didnt take place?")
+          .hasValueGreaterThanOrEqualTo(1);
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
   }
 }