+++ ReleaseNotes.md
@@ -28,6 +28,8 @@
 * TGR-1273: Added support for local port identification of zion servers so that they can be correctly displayed in the
   sequence diagrams.
 * TGR-1086: replace OS specific commands with OSHI library for finding ports being used by external jars.
+* TGR-1292: Tiger-Proxy now supports OCSP-stapling. To enable it, set the configuration key
+  ```tiger.tigerProxy.tls.ocspSigner: myOcspSigner.p12;password``` to the path of the OCSP-signer certificate to be used when generating the OCSP responses.
 * TGR-1282: Traffic Visualization feature no longer experimental. Can be turned on via the configuration:
 
 ```yaml